pyproject.toml
[tool.poetry]
name = "kurs-project"
version = "0.1.0"
description = ""
authors = ["Daniil <danilayukin@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
flake8 = "^7.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^6.0.0b2"
mypy = "^1.11.2"



[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'


[tool.isort]
line_length = 119


[tool.black]
line-length = 119
exclude = '''
(
  /(       # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.mypy_cache
    | \.venv
  )/
)
'''


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
